+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix crash with `Merge Same Dimension Textures` when one material has a non Texture2D assigned and the other material doesn't have any texture assigned in the same slot when they get compared for merging.
 * Fix crash when animator state behaviours fail to clone. [more](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/199)
   * Also logs the type of the behaviour that failed to clone.
+* Fix crash when a vertex in a skinned mesh has 0 weights.
 
 ## v4.5.0
 ### Features